backend/requirements.txt
fastapi
uvicorn
line-bot-sdk
apscheduler
mysql-connector-python
python-dotenv
jaconv
requests
sqlalchemy
pydantic
